Create assingnment3
Implement Circular Queue using Array. Perform following 
 operations on it.
a) Insertion (Enqueue)
b) Deletion (Dequeue)
c) Display

A Circular Queue is an extended version of a normal queue where the last element of the queue is connected to the first element of the queue forming a circle.
It works on the principle of  1st in,1st out.